requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
ipykernel
ipywidgets
torch
pandas
scikit-learn
matplotlib
seaborn
datasets
transformers>=4.50.0
Pillow
trl
wandb
evaluate
accelerate
python-dotenv
peft